third_party/WebKit/LayoutTests/http/tests/devtools/sass/test-ast-editing-2.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/elements-test.js"></script>
<script src="../resources/sass-test.js"></script>
<script src="../resources/editing-test-suite.js"></script>
<style>
pre {
    font-family: monospace;
}
</style>
<script>

function test() {
  TestRunner.evaluateInPagePromise('getCSS()')
      .then(result => SASSTestRunner.parseCSS('', result.value))
      .then(ast => SASSTestRunner.runEditingTests(ast))
      .catch(console.error.bind(console))
      .then(TestRunner.completeTest);
}

function getCSS()
{
    return document.querySelector(".snippet").textContent;
}

</script>
</head>

<body onload="runTest()">
<p>
Verifies AST editing of minified source.
</p>
<pre class="snippet">div{color:red;/*display:block;*/margin:0}</pre>
</body>
</html>
